Moved command processing to a Command Executor
