+++ plugin.yml
@@ -1,5 +1,5 @@
 name: CustomWhitelist
-version: 0.5-RELEASE
+version: 0.6-DEV
 description: Allows modification to the whitelist even when Mojang UUID queries are rejected
 author: Airship Pirates Programmers
 main: org.minecraftairshippirates.customwhitelist.CustomWhitelistPlugin
